tweak file in mockchain
