+++ src/Cooked/MockChain.hs
@@ -7,10 +7,12 @@ import Cooked.MockChain.Balancing as X
 import Cooked.MockChain.Common as X
 import Cooked.MockChain.Error as X
 import Cooked.MockChain.Instances as X
+import Cooked.MockChain.Journal as X
 import Cooked.MockChain.Misc as X
 import Cooked.MockChain.Read as X
 import Cooked.MockChain.State as X
 import Cooked.MockChain.Testing as X
+import Cooked.MockChain.Tweak as X
 import Cooked.MockChain.UtxoSearch as X
 import Cooked.MockChain.UtxoState as X
 import Cooked.MockChain.Write as X
